chore: separate content and bundle scripts
