+++ src/Highlight/contentScript.ts
@@ -1,33 +1,11 @@
-import { RouterliciousDocumentServiceFactory } from "@fluidframework/routerlicious-driver";
-import { getNostrUser } from "../Nostr";
-import {
-	createNostrCreateNewRequest,
-	MockCollabRelay,
-	NostrCollabLoader,
-	NostrRelayTokenProvider,
-	NostrRelayUrlResolver,
-	StaticCodeLoader,
-} from "../Nostrcollab";
-import { HighlightContainerRuntimeFactory } from "./container";
-import {
-	ActionResponse,
-	IHighlight,
-	IHighlightCollection,
-	IHighlightCollectionAppModel,
-	MessageAction,
-	StorageKey,
-} from "./types";
-import {
-	tryReadLocalStorage,
-	sendMessage,
-	sha256Hash,
-	writeLocalStorage,
-	tryWriteLocalStorage,
-} from "./utils";
+import { ActionResponse, MessageAction } from "./types";
 
 let color: string = "FAA99D";
 const HIGHLIGHT_KEY: string = "NPKryv4iXxihMRg2gxRkTfFhwXmNmX9F";
 
+/**
+ * Listen for highlight mesages and take actions that render highlights on the page
+ */
 chrome.runtime.onMessage.addListener(
 	(request: { action: MessageAction; data: any }, _sender, sendResponse) => {
 		let outcome: ActionResponse;
@@ -150,52 +128,3 @@ function getHighlightedMark(): HTMLElement | null {
 	}
 	return parent;
 }
-
-let collab: IHighlightCollection | null = null;
-
-chrome.tabs.onUpdated.addListener(async (_tabId, changeInfo, tab) => {
-	// Load collab whenever on fully loaded tabs, if the tab is active
-	if (changeInfo.status === "complete" && tab.active) {
-		const collabRelayUrl = process.env.COLLAB_RELAY_URL ?? "http://localhost:7070";
-		const collabRelay = new MockCollabRelay("wss://mockcollabrelay", 1, collabRelayUrl);
-
-		const tokenProvider = new NostrRelayTokenProvider(collabRelay, await getNostrUser());
-
-		// Create a new Fluid loader, load the highlight collection
-		const loader = new NostrCollabLoader<IHighlightCollectionAppModel>({
-			urlResolver: new NostrRelayUrlResolver(collabRelay),
-			documentServiceFactory: new RouterliciousDocumentServiceFactory(tokenProvider),
-			codeLoader: new StaticCodeLoader(new HighlightContainerRuntimeFactory()),
-			generateCreateNewRequest: createNostrCreateNewRequest,
-		});
-
-		let storageKey = await sha256Hash(tab.url ?? "");
-		let collabId = await tryReadLocalStorage<string>(storageKey);
-
-		if (!collabId) {
-			const createResponse = await loader.createDetached("0.1.0");
-			collab = createResponse.collab.highlightCollection;
-			tryWriteLocalStorage<string>(storageKey, await createResponse.attach());
-		} else {
-			collab = (await loader.loadExisting(collabId)).highlightCollection;
-		}
-	}
-
-	if (collab !== null) {
-		// Listen for changes to the highlight collection
-		const changeListener = async () => {
-			const highlights = await collab!.getHighlights();
-
-			// Request render highlights on canvas
-			await sendMessage<IHighlight[]>({
-				action: MessageAction.RENDER_HIGHLIGHTS,
-				data: highlights,
-			});
-
-			// Notify render of highlights on popup UI
-			writeLocalStorage<IHighlight[]>(StorageKey.COLLAB_HIGHLIGHTS, highlights);
-		};
-
-		collab.on("highlightCollectionChanged", changeListener);
-	}
-});
